TG-91: change helpers folder to hooks folder
